index.html 분석 그래프 추가, 시세 데이터 반영 코드 수정

+++ src/main/java/meatmeet/meatmeet/service/MemberService.java
@@ -81,7 +81,14 @@ public class MemberService {
 		memberRepository.updateRecipe(recipe);
 	}
 	
-	public void deleteRecipe(Long recipeId) {
-		memberRepository.deleteRecipe(recipeId);
+	public void deleteRecipe(String memberId, Long recipeId) {
+		Optional<Recipe> recipe = findByRecipeId(memberId, recipeId);
+		
+		if(recipe.isPresent()) {
+			String imgUrl = recipe.get().getImgPath();
+			s3Uploader.deleteImgFile(imgUrl);
+			
+			memberRepository.deleteRecipe(recipeId);
+		}
 	}
 }
